Deployment/requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
pillow
plotly
scikit-learn==1.4.1.post1

